Use the official Heroku Ruby buildpack.
For some reason the unofficial JRuby buildpack that Docverter
had been using started breaking. See this stack overflow for
the cure:

http://stackoverflow.com/questions/9078871/sinatra-jruby-heroku-jruby-no-such-file-or-directory-trinidad-loaderror

This commit adds enough metadata to the Gemfile so that we can
start using the official Heroku Ruby buildpack instead of relying
on the unofficial JRuby buildpack.

Fixes #18
Fixes #19